Move all the one-time boilerplate to a single file
This is definitely easier to read than splitting everything into
little sections.  The one-time boilerplate is reasonably small.
We leave the repetitive stuff (methods, signals, etc.) in separate modules.
